backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
transformers>=4.21.0
torch>=1.12.0
pdfplumber>=0.7.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
nltk>=3.7
spacy>=3.4.0